Refactor knowledge hustle modules
Manual testing: harness simulation of study acceptance/completion.
