+++ src/game/hustles/knowledge/constants.js
@@ -1,0 +1,1 @@
+export const STUDY_DIRTY_SECTIONS = Object.freeze(['cards', 'dashboard', 'player']);
